requirement.txt
Django
djangorestframework
djangorestframework-jwt
beautifulsoup4
lxml